requirements.txt
discord.py>=1.4,<1.5
seaborn
lxml
pillow
aiocache
pytest
psycopg2
bs4
aiohttp
